Implemented: max number of cover images in cache. Fixed wrong album arts (previously).

+++ MiniPie.Core/ICoverService.cs
@@ -4,7 +4,7 @@ using MiniPie.Core.SpotifyLocal;
 namespace MiniPie.Core {
     public interface ICoverService {
         Task<string> FetchCover(Status status);
-        double CacheSize();
+        long CacheSize();
         void ClearCache();
     }
 }
